renames util file to format-xml.ts
