+++ src/app/views/common/monaco/util/format-xml.ts
